refactor(agents): extract session-file/hook runtimes and terminal lifecycle seam
